+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Prowide Core - CHANGELOG
 
 #### 9.4.15 - SNAPSHOT
+  * (PW-1812) Updated the narrative resolver, format 2 (used in field 72 for example), to allow empty values as part of the narrative fragment
   * Updated validators for BIC, country, and currency constraints to utilize keywords for i18n-compatible messages
   * Deprecated unnecessary methods in the SafeXmlUtils class
-  * Narrative Fields type 2 support empty values as valid values 
 
 #### 9.4.14 - December 2023
   * (PW-1718) Changed the getComponentLabel(component) in Field59F to be dynamic based on the line identifiers (similar to existing API in Field50F)